Creating a code that can jumping once when grounded.
Player objects that are on a ground object with a "Ground" layer can now jump when the space key is pressed.
Also, if the player object leaves the ground, it will no longer be able to jump when the space key is pressed again unless it returns to the ground.